// AuctionPLL/src/app/generated-models/lot-models/images-view-model.ts
/* eslint-disable */
/**
 * This is a TypeGen auto-generated file.
 * Any changes made to this file can be lost when this file is regenerated.
 */

export class ImagesViewModel {
    id: number;
    image1: string;
    image2: string;
    image3: string;
    image4: string;
    image5: string;
    image6: string;
    image7: string;
    image8: string;
    image9: string;
}